+++ snuba_sdk/metrics_query.py
@@ -4,6 +4,7 @@ from dataclasses import dataclass, replace
 from datetime import datetime
 from typing import Any
 
+from snuba_sdk.expressions import Limit, Offset
 from snuba_sdk.formula import Formula
 from snuba_sdk.metrics_query_visitors import SnQLPrinter, Validator
 from snuba_sdk.query import BaseQuery
@@ -29,6 +30,8 @@ class MetricsQuery(BaseQuery):
     end: datetime | None = None
     rollup: Rollup | None = None
     scope: MetricsScope | None = None
+    limit: Limit | None = None
+    offset: Offset | None = None
 
     def _replace(self, field: str, value: Any) -> MetricsQuery:
         new = replace(self, **{field: value})
@@ -58,6 +61,12 @@ class MetricsQuery(BaseQuery):
         if not isinstance(scope, MetricsScope):
             raise InvalidQueryError("scope must be a MetricsScope")
         return self._replace("scope", scope)
+
+    def set_limit(self, limit: int) -> MetricsQuery:
+        return self._replace("limit", Limit(limit))
+
+    def set_offset(self, offset: int) -> MetricsQuery:
+        return self._replace("offset", Offset(offset))
 
     def validate(self) -> None:
         Validator().visit(self)