+++ CHANGES.rst
@@ -1,6 +1,14 @@
 Changelog and versioning
 ==========================
 
+2.0.8
+-----
+
+- Support the `limit` and `offset` field in the MetricsQuery
+    - This will set the appropriate LIMIT and OFFSET clauses in the resulting SQL query,
+    allowing for pagination
+
+
 2.0.7
 -----
 - Introduce MQL grammar and parser